requirements.txt
python-docx
python-docx-replace
python-dotenv